Updates to testing, requires changes to vscode launch.json with correct env args for each storage-node test, to ensure no nock etc
