+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "test/azure-storage-node"]
-	path = test/azure-storage-node
+	path = externaltests/azure-storage-node
 	url = https://github.com/Azure/azure-storage-node
